pax-app/Wiki#142 add photo profile component
