+++ lib/screens/became_provider_screen/became_provider_screen.dart
@@ -32,34 +32,7 @@ class _BecameProviderScreenState extends State<BecameProviderScreen> {
             Stack(
               alignment: Alignment.bottomCenter,
               children: <Widget>[
-                Container(
-                  child: _photo == null
-                      ? Container()
-                      : Container(
-                          decoration: BoxDecoration(
-                            border: Border.all(
-                                color: Theme.of(context).accentColor,
-                                width: 2.0),
-                            borderRadius:
-                                BorderRadius.all(Radius.circular(15.0)),
-                            color: Colors.transparent,
-                            image: DecorationImage(
-                              image: FileImage(
-                                _photo,
-                              ),
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                        ),
-                  width: 120.0,
-                  height: 120.0,
-                  decoration: BoxDecoration(
-                    border: Border.all(
-                        color: Theme.of(context).accentColor, width: 2.0),
-                    borderRadius: BorderRadius.all(Radius.circular(15.0)),
-                    color: Theme.of(context).accentColor,
-                  ),
-                ),
+                
                 InkWell(
                   onTap: () async {
                     debugPrint(_photo.toString());
